Structure reformed. (#215)
